rename vak/dataset/annot.py to vak/dataset/annotation.py
