+++ src/vak/dataset/__init__.py
@@ -1,5 +1,5 @@
 """module that handles datasets:
 spectrograms made from audio files of vocalizations, and associated annotations"""
 from .classes import MetaSpect, Vocalization, VocalizationDataset
-from . import spect, audio, annot, split
+from . import spect, audio, annotation, split
 from .prep import prep
