+++ tests/unit_tests/test_cli_funcs/test_prep.py
@@ -62,7 +62,7 @@ class TestPrep(unittest.TestCase):
                 if annot_file:
                     annot_list = scribe.to_seq(file=annot_file)
                 else:
-                    annot_files = vak.dataset.annot.files_from_dir(annot_dir=data_dir, annot_format=annot_format)
+                    annot_files = vak.dataset.annotation.files_from_dir(annot_dir=data_dir, annot_format=annot_format)
                     annot_list = scribe.to_seq(file=annot_files)
                 for data_file, annot in zip(data_files_from_dir, annot_list):
                     if set(annot.labels).issubset(labelset):
